+++ user/api.go
@@ -60,6 +60,8 @@ type (
 		DelayBeforeNextLoginAttempt int64 `json:"delayBeforeNextLoginAttempt"`
 		// Maximum number of concurrent login
 		MaxConcurrentLogin int `json:"maxConcurrentLogin"`
+		// Block users to do multiple parallel logins (for load tests we desactivate this)
+		BlockParallelLogin bool `json:blockParallelLogin`
 		//allows for the skipping of verification for testing
 		VerificationSecret string           `json:"verificationSecret"`
 		ClinicDemoUserID   string           `json:"clinicDemoUserId"`